backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pdfplumber
python-docx
pytesseract
pdf2image
Pillow